fix: Fix overflow issue
Because overflow was set to hidden, I've encountered seeing the top of the next icon without seeing it's name or the rest of it. Setting it to auto won't position everything absolute, like what it seems like you guys were trying to mimic, but at least every icon can now be seen.